Change line number parameter to long data type

+++ src/main/scala/http/LineController.scala
@@ -12,7 +12,7 @@ import com.twitter.finatra.request.RouteParam
 import com.twitter.util.Future
 
 case class LineRequest(
-  @RouteParam num: Int
+  @RouteParam num: Long
 )
 
 class LineController @Inject() (
